conky: modify start script
Add conky option -f to force start
Some minor rework

+++ conky/startConky.sh
@@ -1,12 +1,36 @@
 #!/bin/bash
-process_num=$(ps -ef | grep conky | grep -v "grep" | wc -l)
-if [ "$process_num" -gt 1 ]; then
-    echo "Conky's already running"
+
+# set an initial value
+FORCE=false
+CONKY_DIR=$HOME/sources/dotfiles/conky
+
+# read the options, exit if parsing error
+# 'p' or 'path' has required arg
+# 'f' or 'force' has no arg, acting as flag.
+OPTS=$(getopt -o p:f --long path:,force -n 'startConky.sh' -- "$@") || exit 1
+eval set -- "$OPTS"
+
+while true ; do
+    case "$1" in
+        -p|--path)
+            case "$2" in
+                    "")     shift 2 ;;
+                    *)      CONKY_DIR=$2;               shift 2 ;;
+            esac ;;
+        -f|--force)         FORCE=true;                 shift ;;
+        --)                 shift;                      break ;;
+        *)                  echo "Internal error!" ;    exit 1 ;;
+    esac
+done
+
+if $FORCE; then killall conky; fi
+
+if [ "$(pgrep -c conky)" -gt 1 ]; then
+    echo "conky's already running"
 else
-    {
-    killall conky
-    conky --config=$HOME/sources/dotfiles/conky/conkyrc_default &
-    conky --config=$HOME/sources/dotfiles/conky/conkyrc_wordclock &
-    } &>/dev/null
-    echo "Conkys started"
+    #in case one still alive
+    killall conky >/dev/null 2>&1
+    conky --config="$CONKY_DIR"/conkyrc_default --daemonize --quiet &&
+    conky --config="$CONKY_DIR"/conkyrc_wordclock --daemonize --quiet &&
+    echo "conkys started"
 fi
